backend/bot_telegram/requirements.txt
requests==2.31.0
python-telegram-bot==21.3
python-dotenv>=1.0.0
httpx>=0.27.0
